requirements.txt
pydantic
requests
aiogram
python-dotenv